lib/fs, lib/model: Cover more windowsyness sanitizing paths (fixes #7075)
